ref-system/Source code/DD/quanlycongviec-backend/src/requirements.txt
mysql-connector-python
flask
flask_cors
tornado
Flask-Mail
pyotp
Padding
pycryptodomex
pycryptodome
numpy
pandas
requests
opencv-python
img2pdf
pyjwt==1.7.1
networkx
graphviz
openpyxl